writing create or update menu

+++ src/main/java/com/foodorderbe/foodorderbe_artifact/requests/DishMenuCreateReq.java
@@ -2,22 +2,22 @@ package com.foodorderbe.foodorderbe_artifact.requests;
 
 public class DishMenuCreateReq {
     private Long dishId;
-    private Long count;
+    private int count;
     public Long getDishId() {
         return dishId;
     }
     public void setDishId(Long dishId) {
         this.dishId = dishId;
     }
-    public Long getCount() {
+    public int getCount() {
         return count;
     }
-    public void setCount(Long count) {
+    public void setCount(int count) {
         this.count = count;
     }
     public DishMenuCreateReq() {
     }
-    public DishMenuCreateReq(Long dishId, Long count) {
+    public DishMenuCreateReq(Long dishId, int count) {
         this.dishId = dishId;
         this.count = count;
     }
